+++ cypress/integration/unit_tests/add_layer_test.js
@@ -2,7 +2,10 @@ import {ColorStyle, LayerType} from '../../../docs/firebase_layers.js';
 import {processNewEeLayer, processNonEeLayer} from '../../../docs/import/add_layer.js';
 import {disasterData, getCurrentLayers} from '../../../docs/import/manage_layers_lib.js';
 import * as Toast from '../../../docs/toast.js';
-import {createTrs} from '../../support/import_test_util.js';
+import {
+  createTrs, setUpSavingStubs,
+  waitForPromiseAndAssertSaves
+} from '../../support/import_test_util.js';
 import {createAndAppend, setDisasterAndLayers} from '../../support/import_test_util.js';
 import {loadScriptsBeforeForUnitTests} from '../../support/script_loader.js';
 
@@ -10,12 +13,8 @@ const mockAsset = 'mockAsset';
 
 describe('Unit tests for adding layers', () => {
   loadScriptsBeforeForUnitTests('ee', 'firebase', 'jquery');
-  beforeEach(() => {
-    const toastStub = cy.stub(Toast, 'showToastMessage');
-    cy.wrap(toastStub.withArgs('Saving...', -1)).as('savingStub');
-    cy.wrap(toastStub.withArgs('Saved')).as('savedStub');
-    disasterData.clear();
-  });
+  setUpSavingStubs();
+  beforeEach(() => disasterData.clear());
 
   afterEach(() => disasterData.clear());
 
@@ -117,27 +116,6 @@ describe('Unit tests for adding layers', () => {
     });
   });
 });
-
-/**
- * @param {Promise<any>} promise
- * @return {Cypress.Chainable<void>}
- */
-function waitForPromiseAndAssertSaves(promise) {
-  cy.wrap(promise);
-  expectStubCalledOnce('savingStub');
-  return expectStubCalledOnce('savedStub');
-}
-
-/**
- * @param {string} stubName Name of stub, stored as Cypress alias
- * @return {Cypress.Chainable<void>}
- */
-function expectStubCalledOnce(stubName) {
-  return cy.get('@' + stubName).then((/** Sinon.SinonSpy */ stub) => {
-    expect(stub).to.be.calledOnce;
-    stub.resetHistory();
-  });
-}
 
 /**
  * Stubs ee.FeatureCollection with the given number of features
